+++ lib/keycloak-admin/client/client_client.rb
@@ -6,11 +6,26 @@ module KeycloakAdmin
       @realm_client = realm_client
     end
 
+    def save(client_representation)
+      execute_http do
+        RestClient::Resource.new(clients_url, @configuration.rest_client_options).post(
+          client_representation.to_json, headers
+        )
+      end
+    end
+
     def list
       response = execute_http do
         RestClient::Resource.new(clients_url, @configuration.rest_client_options).get(headers)
       end
       JSON.parse(response).map { |client_as_hash| ClientRepresentation.from_hash(client_as_hash) }
+    end
+
+    def get_service_account_user(client_id)
+      response = execute_http do
+        RestClient::Resource.new(service_account_user_url(client_id), @configuration.rest_client_options).get(headers)
+      end
+      UserRepresentation.from_hash(JSON.parse(response))
     end
 
     def clients_url(id=nil)
@@ -19,6 +34,10 @@ module KeycloakAdmin
       else
         "#{@realm_client.realm_admin_url}/clients"
       end
+    end
+
+    def service_account_user_url(client_id)
+      "#{clients_url(client_id)}/service-account-user"
     end
   end
 end